modularize homepage
