+++ frontend/src/components/modules/default/defaultContext.js
@@ -45,7 +45,10 @@ const initState = {
 		q1: '',
 		q2: ''
 	},
+
+	// Homepage
 	recentSearches: [],
+	crawlerSources: [],
 	
 	// Tutorial
 	showTutorial: false,
